requirements.txt
mlflow
notebook
python-box
pyYAML
tqdm
typeguard
joblib
types-PyYAML
Flask
Flask-Cors
pandas
numpy
scikit-learn
scipy
matplotlib
seaborn
dagshub
-e .